rom_quirks: Separate file_contexts injection code
* Use a separate .c file for the 'inject_file_contexts()'
  function which will handle both text and binary format
  file_contexts.

* Only rely on the 'multirom_exclusion_path' string array to
  add the needed exclusions. This will become useful once the
  multirom directory changes.

* This also allows us to use the same file for exclusions in
  TWRP.
